Added the auto-value plugin stub.
The auto-value plugin and its extension has been created. The extension provides access to the plugins configuration and holds some default values. The plugin already provides a task stub that will generate the auto-value sources directory. The plugin sets up the core java plugin and makes the 'compileJava' task depend on the generation of the auto-value sources.